+++ src/candy_types.h
@@ -22,7 +22,6 @@ extern "C"{
 #include "candy_config.h"
 #include <stdint.h>
 #include <stdbool.h>
-#include <stddef.h>
 #include <stdio.h>
 
 #ifdef __GNUC__
@@ -41,7 +40,7 @@ typedef struct candy_pair candy_pair_t;
 typedef struct candy_block candy_block_t;
 typedef struct candy_state candy_state_t;
 
-typedef int (*candy_reader_t)(char *, const size_t, void *);
+typedef int (*candy_reader_t)(char *, const int, void *);
 
 /**
   * @brief c-type function
